@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 
 class BgModel{
 
-  LinearGradient linearGradient;
-  Color hintColor;
-  Color textColor;
+  LinearGradient? linearGradient;
+  Color? hintColor;
+  Color? textColor;
 
   BgModel({this.linearGradient, this.hintColor, this.textColor});
 
