@@ -14,11 +14,11 @@ import 'package:story_picker/story_picker.dart';
 import 'package:video_player/video_player.dart';
 import 'package:visibility_detector/visibility_detector.dart';
 
-Options options;
+Options? options;
 
 class Gallery extends StatelessWidget {
 
-  Gallery(Options opt) { options = opt; }
+  Gallery(Options? opt) { options = opt; }
 
   @override
   Widget build(BuildContext context) {
@@ -36,7 +36,7 @@ class GalleryView extends StatefulWidget {
 }
 
 class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClientMixin{
-  GalleryProvider galleryProvider;
+  late GalleryProvider galleryProvider;
 
   @override
   bool get wantKeepAlive => true;
@@ -48,7 +48,7 @@ class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClient
     galleryProvider =  Provider.of<GalleryProvider>(context, listen: false);
     galleryProvider.getFilesPath();
 
-    galleryProvider.translations = options.translations;
+    galleryProvider.translations = options!.translations;
   }
 
   @override
@@ -60,7 +60,7 @@ class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClient
   Widget build(BuildContext context) {
     super.build(context);
     return Scaffold(
-      backgroundColor: options.customizationOptions.galleryCustomization.bgColor,
+      backgroundColor: options!.customizationOptions.galleryCustomization.bgColor,
       appBar: AppBar(
         elevation: 0.0,
         title: Consumer<GalleryProvider>(
@@ -68,8 +68,8 @@ class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClient
 
               return provider.folders != null ? DropdownButtonHideUnderline(
                   child: DropdownButton<FolderModel>(
-                    items: provider.getItems(),
-                    onChanged: (FolderModel folder) => provider.onFolderSelected(folder),
+                    items: provider.getItems() as List<DropdownMenuItem<FolderModel>>?,
+                    onChanged: (FolderModel? folder) => provider.onFolderSelected(folder!),
                     value: provider.selectedFolder,
                   )
               ) : Container();
@@ -79,7 +79,7 @@ class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClient
         leading: GestureDetector(
           child: Icon(
             Icons.clear,
-            color: options.customizationOptions.galleryCustomization.iconsColor,
+            color: options!.customizationOptions.galleryCustomization.iconsColor,
           ),
           onTap: (){
             Navigator.pop(context, null);
@@ -91,7 +91,7 @@ class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClient
               padding: EdgeInsets.symmetric(horizontal: 16),
               child: Icon(
                 Icons.check,
-                color: options.customizationOptions.galleryCustomization.iconsColor,
+                color: options!.customizationOptions.galleryCustomization.iconsColor,
               ),
             ),
             onTap: (){
@@ -99,7 +99,7 @@ class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClient
             },
           )
         ],
-        backgroundColor: options.customizationOptions.appBarColor,
+        backgroundColor: options!.customizationOptions.appBarColor,
       ),
       body: SafeArea(
           child: Column(
@@ -113,15 +113,15 @@ class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClient
                       Container(
                           height: MediaQuery.of(context).size.height * 0.35,
                           width: MediaQuery.of(context).size.width,
-                          color: options.customizationOptions.galleryCustomization.bgColor,
+                          color: options!.customizationOptions.galleryCustomization.bgColor,
                           child: Stack(
                             children: [
-                              provider.selectedFile.type == AssetType.image
+                              provider.selectedFile!.type == AssetType.image
                                   ? GalleryImagePreview(provider)
-                                  : GalleryVideoPreview(provider, Key(path.basename(provider.selectedFile.filePath))),
-                              options.customizationOptions.galleryCustomization.maxSelectable > 1 ? Positioned(
+                                  : GalleryVideoPreview(provider, Key(path.basename(provider.selectedFile!.filePath!))),
+                              options!.customizationOptions.galleryCustomization.maxSelectable > 1 ? Positioned(
                                   right: 20,
-                                  bottom: provider.selectedFile.type == AssetType.video ? 50 : 5,
+                                  bottom: provider.selectedFile!.type == AssetType.video ? 50 : 5,
                                   child: GestureDetector(
                                     child: Container(
                                       width: 36,
@@ -153,16 +153,16 @@ class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClient
               Consumer<GalleryProvider>(
                   builder: (ctx, provider, child){
 
-                    return provider.selectedFolder != null && provider.selectedFolder.files.length > 0
+                    return provider.selectedFolder != null && provider.selectedFolder!.files!.length > 0
                         ? Container(
-                            color: options.customizationOptions.galleryCustomization.bgColor,
+                            color: options!.customizationOptions.galleryCustomization.bgColor,
                             height: MediaQuery.of(context).size.height * 0.42,
                             child: GridView.builder(
                                 gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
                                     crossAxisCount: 4, crossAxisSpacing: 4, mainAxisSpacing: 4
                                 ),
                                 itemBuilder: (_, i) {
-                                  var file = provider.selectedFolder.files[i];
+                                  var file = provider.selectedFolder!.files![i];
 
                                   return GalleryItem(
                                     file: file,
@@ -170,7 +170,7 @@ class _GalleryViewState extends State<GalleryView> with AutomaticKeepAliveClient
                                   );
 
                                 },
-                                itemCount: provider.selectedFolder.files.length
+                                itemCount: provider.selectedFolder!.files!.length
                             ),
                           ) : Container();
 
@@ -191,8 +191,8 @@ class GalleryImagePreview extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return PhotoView(
-      imageProvider: FileImage(File(provider.selectedFile.filePath)),
-      backgroundDecoration: BoxDecoration(color: options.customizationOptions.galleryCustomization.bgColor),
+      imageProvider: FileImage(File(provider.selectedFile!.filePath!)),
+      backgroundDecoration: BoxDecoration(color: options!.customizationOptions.galleryCustomization.bgColor),
     );
   }
 }
@@ -209,15 +209,15 @@ class GalleryVideoPreview extends StatefulWidget {
 
 class _GalleryVideoPreviewState extends State<GalleryVideoPreview> {
 
-  Future<bool> _init;
-  ChewieController _chewieController;
-  VideoPlayerController _videoPlayerController;
+  Future<bool>? _init;
+  late ChewieController _chewieController;
+  late VideoPlayerController _videoPlayerController;
 
   @override
   void initState() {
     super.initState();
 
-    _videoPlayerController = VideoPlayerController.file(File(widget.provider.selectedFile.filePath));
+    _videoPlayerController = VideoPlayerController.file(File(widget.provider.selectedFile!.filePath!));
 
     _init = initVideoPlayer();
   }
@@ -252,7 +252,7 @@ class _GalleryVideoPreviewState extends State<GalleryVideoPreview> {
 
         if (!snp.hasData) return Center(
             child: CircularProgressIndicator(
-                backgroundColor: options.customizationOptions.accentColor
+                backgroundColor: options!.customizationOptions.accentColor
             )
         );
 
@@ -275,10 +275,10 @@ class _GalleryVideoPreviewState extends State<GalleryVideoPreview> {
 }
 
 class GalleryItem extends StatefulWidget {
-  final FileModel file;
-  final GalleryProvider provider;
+  final FileModel? file;
+  final GalleryProvider? provider;
 
-  GalleryItem({Key key, this.file, this.provider}) : super(key: key);
+  GalleryItem({Key? key, this.file, this.provider}) : super(key: key);
 
   @override
   _GalleryItemState createState() => _GalleryItemState();
@@ -307,27 +307,27 @@ class _GalleryItemState extends State<GalleryItem> with AutomaticKeepAliveClient
       fit: StackFit.expand,
       children: [
         GestureDetector(
-          child: File(widget.file.thumbPath) != null
+          child: File(widget.file!.thumbPath!) != null
               ? Image.file(
-            File(widget.file.thumbPath),
+            File(widget.file!.thumbPath!),
             fit: BoxFit.cover,
           )
               : Container(color: Colors.grey),
           onTap: () {
-            widget.provider.selectedFile = widget.file;
-            if (widget.provider.multiSelect) widget.provider.toggleCheckState(widget.file);
+            widget.provider!.selectedFile = widget.file;
+            if (widget.provider!.multiSelect) widget.provider!.toggleCheckState(widget.file);
           },
           onLongPress: (){
 
-            if (options.customizationOptions.galleryCustomization.maxSelectable == 1) return;
+            if (options!.customizationOptions.galleryCustomization.maxSelectable == 1) return;
 
-            widget.provider.multiSelect = !widget.provider.multiSelect;
-            widget.provider.toggleCheckState(widget.file);
-            widget.provider.selectedFile = widget.file;
+            widget.provider!.multiSelect = !widget.provider!.multiSelect;
+            widget.provider!.toggleCheckState(widget.file);
+            widget.provider!.selectedFile = widget.file;
 
           },
         ),
-        widget.provider.multiSelect ? Positioned(
+        widget.provider!.multiSelect ? Positioned(
             top: 5,
             right: 5,
             child: GestureDetector(
@@ -336,12 +336,12 @@ class _GalleryItemState extends State<GalleryItem> with AutomaticKeepAliveClient
                 height: 24,
                 padding: EdgeInsets.only(top: 2),
                 decoration: new BoxDecoration(
-                    color: widget.provider.getCheckState(widget.file) ? options.customizationOptions.accentColor : Colors.white70,
+                    color: widget.provider!.getCheckState(widget.file) ? options!.customizationOptions.accentColor : Colors.white70,
                     shape: BoxShape.circle,
-                    border: Border.all(width: 1.5, color: options.customizationOptions.galleryCustomization.bgColor)
+                    border: Border.all(width: 1.5, color: options!.customizationOptions.galleryCustomization.bgColor)
                 ),
-                child: widget.provider.getCheckState(widget.file)
-                    ? Text(widget.provider.getCheckNumber(widget.file).toString(),
+                child: widget.provider!.getCheckState(widget.file)
+                    ? Text(widget.provider!.getCheckNumber(widget.file).toString(),
                     style: TextStyle(
                         color: Colors.white
                     ),
@@ -350,16 +350,16 @@ class _GalleryItemState extends State<GalleryItem> with AutomaticKeepAliveClient
               ),
               onTap: (){
 
-                widget.provider.toggleCheckState(widget.file);
-                widget.provider.selectedFile = widget.file;
+                widget.provider!.toggleCheckState(widget.file);
+                widget.provider!.selectedFile = widget.file;
 
               },
             )
         ) : Container(),
-        widget.file.duration != null && widget.file.type == AssetType.video ? Positioned(
+        widget.file!.duration != null && widget.file!.type == AssetType.video ? Positioned(
             right: 5,
             bottom: 5,
-            child: Text(StoryUtils.printDuration(widget.file.duration), style: TextStyle(color: Colors.white),)
+            child: Text(StoryUtils.printDuration(widget.file!.duration!), style: TextStyle(color: Colors.white),)
         ) : Container()
       ],
     ) : Container();
